Convert all runtime endpoints to DIDs
This changeset introduces DIDs to all Polymesh runtime modules. DIDs
serve as an additional identity layer with ability to gather multiple
users under a single identifier. A notable API change is adding an
additional "did" field to most endpoints and constraining users to act
only on behalf of a DID they partake in.

* Remove all appearances of AccountId in storage and APIs (with few
exceptions)
* Add DID authentication checks to most endpoints
* Adjust for unit tests
* Disable external YAML asset test cases (it should be less
labor-intensive to develop them when the API is more mature)